Add and Remove with API
